requirements/local.txt
-r base.txt
-e .
-e ../base
-e ../finance
-e ../login
-e ../mail
-e ../stock
beautifulsoup4
devpi-client
django_extensions
factory-boy
gitpython
hgapi
ipdb
psycopg2
pytest-cov
pytest-django
pytest-flakes
pytest-pep8
pyyaml
walkdir
django-debug-toolbar
